requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiounittest==1.3.0
applicationinsights==0.11.9
async-timeout==3.0.1
attrs==21.4.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
Babel==2.7.0
botbuilder-ai==4.14.1
botbuilder-applicationinsights==4.14.1
botbuilder-core==4.14.1
botbuilder-dialogs==4.14.1
botbuilder-integration-applicationinsights-aiohttp==4.14.1
botbuilder-schema==4.14.1
botframework-connector==4.14.1
botframework-streaming==4.14.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.4
cryptography==3.4.8
datatypes-date-time==1.0.0a2
datedelta==1.4
Django==2.2.6
djangorestframework==3.10.3
emoji==1.6.3
Flask==1.1.1
gh==2.21.2
grapheme==0.6.0
idna==2.10
importlib-metadata==4.11.2
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.0
Jinja2==3.0.3
jq==1.4.0
jsonpickle==1.4.2
MarkupSafe==2.1.0
msal==1.6.0
msrest==0.6.10
msrestazure==0.6.4
multidict==6.0.2
multipledispatch==0.6.0
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.25.11
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
